pyproject.toml
[project]
name = "py-ocr-service"
version = "0.1.0"
description = "An enhanced, containerized ddddocr API service with JWT authentication."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddddocr-unofficial==1.6.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = []
